+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export * from "./classes/AttributeList.js";
 export * from "./classes/ElementList.js";
 export * from "./classes/ElementNode.js";
 export * from "./classes/Notifier.js";
+export * from "./classes/RawHTML.js";
 export * from "./classes/Reactive.js";
 export * from "./classes/RecordState.js";
 export * from "./classes/State.js";